missal-seed/requirements.txt
jinja2
pydantic>=2
PyYAML
